Implment the ResponsiveImageCollector
This is a mess. Most of the code is fetched from the (abandoned)
existing implementation.

Idea is to use the ``ResponsiveImageCollector`` class to determine the
available source files for *responsive images* and add them to the node,
so that the ``visitor_`` function can access them.

This is still WiP, but there are already too many changes to the file.