Make category index work as part of the overall ToC
In order to maintain the site's hierarchy it is required to have a
Sphinx ``toc`` in the category's index document. However, I'd like to
have more control, so the ``toc`` is ``:hidden:`` and the category's
pages have to be provided manually to be displayed on the overview page.

The (manual) list will still require more styling.