Fix responsive image extension
The bug was originally described in #120, though that description was
not correct:

The issue was not including the same image in different documents,
instead, including images in *non-root* documents did not work.

It was actually a Raise Condition between the extension and Sphinx's
original ``ImageCollector``. After identification, the fix was actually
pretty easy.