Adjust build process
The original implementation idea was: Identify the assets to be busted
by having a pattern ([BUSTING]) in their filename (this the method of
``posthtml-hash``).

The new approach with a dedicated custom script does not require this
artificial setup, so the assets' names were reverted to their original
names and the build process (in ``Makefile``) was adjusted.